+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ## [2.7.3] - 2019-05-01
 ### Changed
 - arg_parser.arg_parse2:  Replaced multi-list code section with call to _parse_multi function.
+- arg_parser.arg_parse2:  Replaced single-list code section with call to _parse_single function.
 - arg_parser.arg_file_chk:  Replaced exception code with call to _create_file function.
 
 ### Added